hackathon-dapr/harvester-insights-agent/requirements.txt
# Dapr Agents Framework

# Dapr SDK for pub/sub and service invocation
dapr>=1.12.0

# FastAPI and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# HTTP client for MCP and web requests
httpx>=0.25.0
aiohttp>=3.9.0

# Environment and configuration
python-dotenv>=1.0.0

# Cloud events for Dapr pub/sub
cloudevents>=1.9.0

# File upload support
python-multipart>=0.0.6

# JSON handling
orjson>=3.9.0

# Logging and monitoring
structlog>=23.2.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Note: PostgreSQL access will be through MCP tools, not direct connection
